Update `composer.lock`
As we have added a new package via composer, we need to include the reflected changes in `composer.lock`. This includes a few minor version bumps for other packages, which is both unavoidable and preferable.

Composer also updated the syntax of `web/config.php`, but the changes make it _more_ backwards compatible so I'm happy to keep them.